requirements.txt
opencv-python-headless==4.6.0.66
tensorflow-cpu
keras
pytesseract==0.3.7
fastapi[all]
gunicorn
sklearn